+++ tasks/kmt.py
@@ -1,12 +1,14 @@
-import os
-import filecmp
-import libvirt
-import platform
 from invoke import task
-from invoke.exceptions import Exit
 from glob import glob
+from .kernel_matrix_testing.init_kmt import (
+    KMT_STACKS_DIR,
+    update_kernel_packages,
+    update_rootfs,
+    revert_rootfs,
+    init_kernel_matrix_testing_system,
+    revert_kernel_packages,
+)
 from .kernel_matrix_testing import vmconfig
-from .kernel_matrix_testing.init_kmt import check_and_get_stack
 from .kernel_matrix_testing import stacks
 
 
@@ -26,69 +28,6 @@ def create_stack(ctx, stack=None, branch=False):
 )
 def gen_config(ctx, stack=None, branch=False, vms="", init_stack=False, vcpu="4", memory="8192", new=False):
     vmconfig.gen_config(ctx, stack, branch, vms, init_stack, vcpu, memory, new)
-    
-
-
-def update_kernel_packages(ctx):
-    arch = archs_mapping[platform.machine()]
-    kernel_packages_sum = f"kernel-packages-{arch}.sum"
-    kernel_packages_tar = f"kernel-packages-{arch}.tar"
-
-    ctx.run(
-        f"wget -q https://dd-agent-omnibus.s3.amazonaws.com/kernel-version-testing/{kernel_packages_sum} -O /tmp/{kernel_packages_sum}"
-    )
-
-    current_sum_file = f"{KMT_PACKAGES_DIR}/{kernel_packages_sum}"
-    if filecmp.cmp(current_sum_file, f"/tmp/{kernel_packages_sum}"):
-        print("[-] No update required for custom kernel packages")
-
-    # backup kernel-packges
-    karch = karch_mapping[archs_mapping[platform.machine()]]
-    ctx.run(
-        f"find {KMT_PACKAGES_DIR} -name \"kernel-*.{karch}.pkg.tar.gz\" -type f | rev | cut -d '/' -f 1  | rev > /tmp/package.ls"
-    )
-    ctx.run(f"cd {KMT_PACKAGES_DIR} && tar -cf {kernel_packages_tar} -T /tmp/package.ls")
-    ctx.run(f"cp {KMT_PACKAGES_DIR}/{kernel_packages_tar} {KMT_BACKUP_DIR}")
-    ctx.run(f"cp {current_sum_file} {KMT_BACKUP_DIR}")
-    print("[+] Backed up current packages")
-
-    # clean kernel packages directory
-    ctx.run(f"rm -f {KMT_PACKAGES_DIR}/*")
-
-    download_kernel_packages(ctx, revert=True)
-
-    print("[+] Kernel packages successfully updated")
-
-
-
-def update_rootfs(ctx):
-    arch = archs_mapping[platform.machine()]
-    if arch == "x86_64":
-        rootfs = "rootfs-amd64"
-    elif arch == "arm64":
-        rootfs = "rootfs-arm64"
-    else:
-        Exit(f"Unsupported arch detected {arch}")
-
-    ctx.run(
-        f"wget -q https://dd-agent-omnibus.s3.amazonaws.com/kernel-version-testing/{rootfs}.sum -O /tmp/{rootfs}.sum"
-    )
-
-    current_sum_file = f"{KMT_ROOTFS_DIR}/{rootfs}.sum"
-    if filecmp.cmp(current_sum_file, "/tmp/{rootfs}.sum"):
-        print("[-] No update required for root filesystems and bootable images")
-
-    # backup rootfs
-    ctx.run("cp {KMT_ROOTFS_DIR}/{rootfs}.tar.gz {KMT_BACKUP_DIR}")
-    ctx.run("cp {KMT_ROOTFS_DIR}/{rootfs}.sum {KMT_BACKUP_DIR}")
-    print("[+] Backed up rootfs")
-
-    # clean rootfs directory
-    ctx.run(f"rm -f {KMT_ROOTFS_DIR}/*")
-
-    download_rootfs(ctx, revert=True)
-
-    print("[+] Root filesystem and bootables images updated")
 
 
 @task
@@ -99,7 +38,7 @@ def update_resources(ctx):
         return
 
     for stack in glob(f"{KMT_STACKS_DIR}/*"):
-        destroy_stack(ctx, stack=stack)
+        destroy_stack(ctx, stack=stack, force=True)
 
     update_kernel_packages(ctx)
     update_rootfs(ctx)
@@ -113,13 +52,12 @@ def revert_resources(ctx):
         return
 
     for stack in glob(f"{KMT_STACKS_DIR}/*"):
-        destroy_stack(ctx, stack=stack)
+        destroy_stack(ctx, stack=stack, force=True)
 
     revert_kernel_packages(ctx)
     revert_rootfs(ctx)
 
     print("[+] Reverted successfully")
-
 
 
 @task
@@ -128,10 +66,12 @@ def launch_stack(
 ):
     stacks.launch_stack(ctx, stack, branch, ssh_key, x86_ami, arm_ami)
 
+
 @task
 def destroy_stack(ctx, stack=None, branch=False, force=False, ssh_key=""):
     stacks.destroy_stack(ctx, stack, branch, force, ssh_key)
 
+
 @task
-def init_stack(ctx):
-    init_kmt.init_kernel_matrix_testing_system(ctx)
+def init(ctx):
+    init_kernel_matrix_testing_system(ctx)
